+++ src/encoded/static/components/util/hooks.js
@@ -2,7 +2,8 @@
 import { useState, useEffect } from 'react';
 import { ajax } from '@hms-dbmi-bgm/shared-portal-components/es/components/util';
 
-const defaultDownloadAccessObject = {
+// Default download access object for non-logged in users
+export const defaultDownloadAccessObject = {
     open: false,
     'open-early': false,
     'open-network': false,
@@ -10,16 +11,17 @@ const defaultDownloadAccessObject = {
     'protected-early': false,
     'protected-network': false,
     released: false,
-    uploading: false,
-    uploaded: false,
-    retracted: false,
-    'upload failed': false,
-    'to be uploaded by workflow': false,
-    'in review': false,
-    obsolete: false,
-    archived: false,
-    deleted: false,
+    uploading: true,
+    uploaded: true,
+    retracted: true,
+    'upload failed': true,
+    'to be uploaded by workflow': true,
+    'in review': true,
+    obsolete: true,
+    archived: true,
+    deleted: true,
 };
+
 /**
  * Checks the session-properties endpoint to determine the statuses that a user
  * has access to. Ultimately used to determine whether to disable the download
@@ -37,86 +39,51 @@ export const useUserDownloadAccess = (session) => {
 
     useEffect(() => {
         let cancelled = false;
-        setIsAccessResolved(false);
 
-        if (session) {
-            const userDownloadAccessObj = { ...defaultDownloadAccessObject };
-
-            // If session exists, user has access to the following statuses
-            userDownloadAccessObj['open'] = true;
-
-            // Default to true when user is logged in. If it is visible, user
-            // can likely download. Otherwise let backend enforce download
-            // access.
-            userDownloadAccessObj['released'] = true;
-            userDownloadAccessObj['uploading'] = true;
-            userDownloadAccessObj['uploaded'] = true;
-            userDownloadAccessObj['retracted'] = true;
-            userDownloadAccessObj['upload failed'] = true;
-            userDownloadAccessObj['to be uploaded by workflow'] = true;
-            userDownloadAccessObj['in review'] = true;
-            userDownloadAccessObj['obsolete'] = true;
-            userDownloadAccessObj['archived'] = true;
-            userDownloadAccessObj['deleted'] = true;
-
-            ajax.load(
-                '/session-properties',
-                (resp) => {
-                    if (cancelled) return; // ignore stale response
-
-                    // Get consortia associated with user
-                    const userConsortia = resp?.details?.consortia || [];
-
-                    // Check if user is a member of SMaHT consortium
-                    const smaht_uuid = '358aed10-9b9d-4e26-ab84-4bd162da182b'; // SMaHT consortium UUID
-                    const isMember = userConsortia?.includes(smaht_uuid);
-
-                    // Get groups associated with user
-                    const userGroups = resp?.details?.groups || [];
-
-                    if (isMember) {
-                        // User is a member of SMaHT
-                        userDownloadAccessObj['open-early'] = true;
-                        userDownloadAccessObj['open-network'] = true;
-
-                        // User is either admin or dbgap member of SMaHT
-                        if (
-                            userGroups?.includes('admin') ||
-                            userGroups?.includes('dbgap')
-                        ) {
-                            userDownloadAccessObj['protected'] = true;
-                            userDownloadAccessObj['protected-early'] = true;
-                            userDownloadAccessObj['protected-network'] = true;
-                        }
-                    } else {
-                        // User is not a member of SMaHT
-                        if (userGroups?.includes('public-dbgap')) {
-                            userDownloadAccessObj['protected'] = true;
-                        }
-                    }
-
-                    setDownloadAccessObject(userDownloadAccessObj);
-                    setIsAccessResolved(true);
-                },
-                'GET',
-                (err) => {
-                    if (
-                        err?.notification !==
-                        'No session property information found.'
-                    ) {
-                        console.error(
-                            'ERROR determining session property information:',
-                            err
-                        );
-                        setDownloadAccessObject(defaultDownloadAccessObject);
-                        setIsAccessResolved(true);
-                    }
-                }
-            );
-        } else {
+        // If no session, return default access object
+        if (!session) {
             setDownloadAccessObject(defaultDownloadAccessObject);
             setIsAccessResolved(true);
         }
+
+        setIsAccessResolved(false);
+
+        const baseAccessPerms = {
+            ...defaultDownloadAccessObject,
+            open: true, // Logged in users have access to open files
+        };
+
+        ajax.load(
+            '/session-properties',
+            (resp) => {
+                // ignore stale response
+                if (cancelled) return;
+
+                const mergedAccessPerms = {
+                    ...baseAccessPerms,
+                    ...(resp?.download_perms || {}),
+                };
+                console.log('Download Access Object', mergedAccessPerms);
+
+                setDownloadAccessObject(mergedAccessPerms);
+                setIsAccessResolved(true);
+            },
+            'GET',
+            (err) => {
+                if (cancelled) return;
+                if (
+                    err?.notification !==
+                    'No session property information found.'
+                ) {
+                    console.error(
+                        'ERROR determining session property information:',
+                        err
+                    );
+                }
+                setDownloadAccessObject(defaultDownloadAccessObject);
+                setIsAccessResolved(true);
+            }
+        );
 
         return () => {
             cancelled = true;